## Pushing in the final commit for version 1.7.0.
Version 1.7.0 contains multiple enhancements to get around various VPN issues that resulted in the GUI being not accessible when outside the main computer running the deluge instance.

- Fixed logs being stripped out when running as a daemon.
- Added more clear logs that inform how the container is running, such as if it's using environment variables or the config file to load the VPN details.
- Removed some redundant VPN configurations that were left over from the older OpenVPN service. up.py will probably be removed in a later version.
- Added enhancements to the first setup experience for users using the container. If it detects that no config exists, the container will create the necessary config details providing a default downloads path pointing to /downloads and pointing the output interface to tun0 to ensure that there are no chances that data is exposed over the internet outside of the VPN if a disconnection occurs.

Please note, if moving from an older version to this version, you may run into issues with your config files. In earlier versions, the config folder was running as the root user type. Now it runs as the user 1000:1000. This can be modified by changing DEL_GID and DEL_UID environment variables however the container will no longer run as a root user.