Hotfix for broken resolv.conf on container restart
Pushed in changes to shut down the Windscribe service before the container is allowed to close. This should allow the resolv.conf and iptables to revert back to their original configuration. If this process does not happen, the next time the container is resumed, DNS requests will fail to resolve.